requirements.txt
aiohttp
pandas
fastapi
sqlalchemy
uvicorn
pytest
python-dotenv
